FiniteElementAnalysis

Documentation for FiniteElementAnalysis.

FiniteElementAnalysis.axialstresses

FiniteElementAnalysis.globalstiffnessmatrix

FiniteElementAnalysis.prepareelements_beam

FiniteElementAnalysis.prepareelements_line

FiniteElementAnalysis.reducedglobalstiffnessmatrix

FiniteElementAnalysis.shearstresses

FiniteElementAnalysis.solve

FiniteElementAnalysis.axialstresses — Method

axialstresses(elements::Dict{String, LineElements}, nodedisplacements::Vector{Float64}, dims=1) axialstresses(elements::Dict{String, BeamElements}, bendingmoments::Vector{Float64}, distancesfromneutralaxis::Union{<:Real, Vector{<:Real}}, dims=2)

Determine the axial stresses present in element-type system.

If line elements, of two-force members in their local coordinate systems. If beam elements, of beam elements which assumes no direct, axial forces.

See also: prepareelements_line, prepareelements_beam, solve, and transformationmatrixstar.

Arguments

elements: the prepared dictionary according to element type.

nodedisplacements: the vector for displacement of each node in the global coordinate system.

bendingmoments: the vector of reaction moments at each node from solve.F.

distancesfromneutralaxis: the maximum distance(s) of nodes from neutral axis.

dims=1: the number of dimensions being analyzed.

Examples

julia> A, E, L = 1, 10e6, 100 # [in², psi, in]
julia> angles = [120, 0, 210] # [degrees]
julia> L = L ./ abs.(cosd.(angles)) # [in]
julia> elements = Dict("1"=>(1, 2), "2"=>(1, 3), "3"=>(1, 4))
julia> preparedelements = prepareelements_line(elements, A, E, L, angles, dims=2)
julia> nodeboundaryconditions = [2, 3, 4]
julia> F_applied = [(1, (1e3, 1e3))] # (node, (force_x [lb], force_y [lb]))
julia> U = solve(preparedelements, nodeboundaryconditions, F_applied, dims=2).U
julia> axialstresses(preparedelements, U, dims=2) # [psi]
3-element Vector{Float64}:
 -577.3502691896259
 -422.6497308103743
 1000.0000000000001

julia> A, E, I, L = 0.004887968, 200e9, 102e6 * (1/1e3)^4, 6                # [m², Pa, m⁴, m]
julia> Q, c, t_w = 323473.474 * (1/1e3)^3, 176.5e-3, 6.48e-3                # [m³, m, m]
julia> W(x) = -10e3(x) # [N/m]
julia> F_applied = ([(W, 1, 2, 3, 4)])
julia> x = 0:L:18
julia> elements = Dict{String, Tuple{Int, Int}}()
julia> nodeboundaryconditions = []
julia> for node ∈ 1:1:numberofelements+1
julia> if node == 1
julia> push!(nodeboundaryconditions, node)
julia> else
julia> elements["$(node-1)"] = (node-1, node)
julia> if x[node]%L == 0
julia> push!(nodeboundaryconditions, (node, (Inf, 0)))
julia> end
julia> end
julia> end
julia> preparedelements = prepareelements_beam(elements, E, I, L)
julia> sol = solve(preparedelements, nodeboundaryconditions, F_applied)     # [m]
julia> V = [sol.F[2i - 1] for i ∈ 1:1:length(sol.F)÷2]                      # [N]
julia> M = [sol.F[2i] for i ∈ 1:1:length(sol.F)÷2]                          # [N-m]
julia> maximum(axialstresses(preparedelements, M, c))                       # [Pa]
519117.6470587983

FiniteElementAnalysis.globalstiffnessmatrix — Method

globalstiffnessmatrix(elements::Dict{String, LineElements}, dims::Integer=1)::Matrix{Float64} globalstiffnessmatrix(elements::Dict{String, BeamElements}, dims::Integer=2)::Matrix{Float64}

Assembles the global stiffness matrix from the dictionary, elements.

See also: prepareelements_line, prepareelements_beam, and transformationmatrix.

Arguments

elements: entries infer localstiffnessmatrices and nodeconnections for each element in local coordinate system transformed by angle.

dims: the number of dimensions being analyzed.

Examples

julia> A = [0.75, 0.5, 1]   # [in²]
julia> E = 10.6e6           # [psi]
julia> L = [12, 9, 8]       # [in]
julia> elements = Dict("1"=>(1, 2), "2"=>(2, 3), "3"=>(3, 4))
julia> globalstiffnessmatrix(prepareelements_line(elements, A, E, L))
4×4 Matrix{Float64}:
  662500.0  -662500.0         0.0         0.0
 -662500.0        1.25139e6  -5.88889e5   0.0
       0.0       -5.88889e5   1.91389e6  -1.325e6
       0.0        0.0        -1.325e6     1.325e6

julia> E, I, L = 30e6, 510, 60 # [psi, in⁴, in]
julia> W(x) = -12\1e3(x)
julia> F_applied = ([(W, 1, 2, 3)])
julia> elements = Dict("1"=>(1, 2), "2"=>(2, 3))
julia> globalstiffnessmatrix(prepareelements_beam(elements, E, I, L))
6×6 Matrix{Float64}:
  850000.0      2.55e7  -850000.0      2.55e7        0.0      0.0
       2.55e7   1.02e9       -2.55e7   5.1e8         0.0      0.0
 -850000.0     -2.55e7        1.7e6    0.0     -850000.0      2.55e7
       2.55e7   5.1e8         0.0      2.04e9       -2.55e7   5.1e8
       0.0      0.0     -850000.0     -2.55e7   850000.0     -2.55e7
       0.0      0.0           2.55e7   5.1e8        -2.55e7   1.02e9

FiniteElementAnalysis.prepareelements_beam — Function

prepareelements_beam(elements, elementmoduli, elementareasofinertia, elementlengths, nodeangles=0, isdegrees=true; dims=2, stiffnesscoefficients=nothing)::Dict{String, BeamElements}

Return the dictionary with properties specific to those beam elements between node pairs.

See also: globalstiffnessmatrix, solve, and axialstresses.

Arguments

elements::Dict{String, Tuple{Int, Int}}: the dictionary of "element"=>(node 1, node 2) for the element between node pairs.

elementmoduli::Union{<:Real, Vector{<:Real}}: the elastic modulus of each element.

elementareasofinertia::Union{<:Real, Vector{<:Real}}: the cross-sectional area of each element.

elementlengths::Union{<:Real, Vector{<:Real}}: the length of each element.

nodeangles::Union{<:Real, Vector{<:Real}}=0: the angle between the global and local coordinate systems of each element.

isdegrees::Bool=true: control value whether elements of nodeangles is degrees (true) or radians (false).

dims::Integer=2: the number of dimensions being analyzed.

stiffnesscoefficients::Union{<:Real, Vector{<:Real}}: the stiffness coefficient, $k$ associated with each element.

Examples

julia> E, I, L = 30e6, 200, 20         # [psi, in⁴, ft]
julia> elements = Dict{String, Tuple{Int, Int}}("1"=>(1, 2),"2"=>(2, 3))
julia> prepareelements_beam(elements, E, I, L)
Dict{String, FiniteElementAnalysis.BeamElements} with 2 entries:
  "1" => BeamElements((1, 2), 3.0e7, 200.0, 20.0, 750000.0, 0.0, [9.0e6 9.0e7 -9.0e6 9.0e7; 9.0e7 1.2e9 -9.0e7 6.0e8; -9.0e6 -9.0e7 9.0e6 -9.0e7; 9.0e7 6.0e8 -9.0e7 1.2e9])
  "2" => BeamElements((2, 3), 3.0e7, 200.0, 20.0, 750000.0, 0.0, [9.0e6 9.0e7 -9.0e6 9.0e7; 9.0e7 1.2e9 -9.0e7 6.0e8; -9.0e6 -9.0e7 9.0e6 -9.0e7; 9.0e7 6.0e8 -9.0e7 1.2e9])

FiniteElementAnalysis.prepareelements_line — Function

prepareelements_line(elements, elementareas, elementmoduli, elementlengths, nodeangles=0, isdegrees=true; dims=1, stiffnesscoefficients=nothing)::Dict{String, LineElements}

Return the dictionary with properties specific to those line elements between node pairs.

See also: globalstiffnessmatrix, solve, and axialstresses.

Arguments

elements::Dict{String, Tuple{Int, Int}}: the dictionary of "element"=>(node 1, node 2) for the element between node pairs.

elementareas::Union{<:Real, Vector{<:Real}}: the cross-sectional area(s) of elements.

elementmoduli::Union{<:Real, Vector{<:Real}}: the elastic modul(us/i) of elements.

elementlengths::Union{<:Real, Vector{<:Real}}: the length(s) of elements.

nodeangles::Union{<:Real, Vector{<:Real}}=0: the angle between the global and local coordinate systems of elements.

isdegrees::Bool=true: control value whether elements of nodeangles is degrees (true) or radians (false).

dims::Integer=1: the number of dimensions being analyzed.

stiffnesscoefficients::Union{<:Real, Vector{<:Real}}: the stiffness coefficient, $k$ associated with each element.

Examples

julia> A = [0.75, 0.5, 1]   # [in²]
julia> E = 10.6e6           # [psi]
julia> L = [12, 9, 8]       # [in]
julia> elements = Dict("1"=>(1, 2), "2"=>(2, 3), "3"=>(3, 4))
julia> prepareelements_line(elements, A, E, L)
Dict{String, FiniteElementAnalysis.LineElements} with 3 entries:
  "1" => LineElements((1, 2), 0.75, 1.06e7, 12.0, 662500.0, 0.0, [662500.0 -662500.0; -662500.0 662500.0])
  "2" => LineElements((2, 3), 0.5, 1.06e7, 9.0, 5.88889e5, 0.0, [5.88889e5 -5.88889e5; -5.88889e5 5.88889e5])
  "3" => LineElements((3, 4), 1.0, 1.06e7, 8.0, 1.325e6, 0.0, [1.325e6 -1.325e6; -1.325e6 1.325e6])

FiniteElementAnalysis.reducedglobalstiffnessmatrix — Method

reducedglobalstiffnessmatrix(K::Matrix{Float64}, nodes::Union{Vector{Integer}, Vector{Tuple{Integer, Integer}}}, dims::Integer=1)::Matrix{Float64}

Eliminate the rows and columns of K according to nodes which implies the displacement of the appropriate node is known: e.g. is fixed or a prescribed displacement.

See also: globalstiffnessmatrix.

Examples

julia> A = [0.75, 0.5, 1]   # [in²]
julia> E = 10.6e6           # [psi]
julia> L = [12, 9, 8]       # [in]
julia> elements = Dict("1"=>(1, 2), "2"=>(2, 3), "3"=>(3, 4))
julia> K = globalstiffnessmatrix(prepareelements_line(elements, A, E, L))
julia> skipnodes = Vector{Integer}([1, 4])
julia> reducedglobalstiffnessmatrix(K, skipnodes)
2×2 Matrix{Float64}:
  1.25139e6  -5.88889e5
 -5.88889e5   1.91389e6

FiniteElementAnalysis.shearstresses — Method

shearstresses(elements, shearforces, firstmomentofareas, bendingthicknesses, dims=2)

Determine the axial stresses present in beam element system.

See also: prepareelements_beam, solve, and transformationmatrixstar.

Arguments

elements::Dict{String, BeamElements}: the prepared dictionary according to element type.

shearforces::Vector{Float64}: the vector for reaction, shear forces at each node in the global coordinate system.

firstmomentofarea::Union{<:Real, Vector{<:Real}}: the first moment of area(s), $Q$ above the neutral axis.

bendingthicknesses::Union{<:Real, Vector{<:Real}}: the web thickness of I-beam.

dims::Integer=2: the number of dimensions being analyzed.

Examples

julia> A, E, I, L = 0.004887968, 200e9, 102e6 * (1/1e3)^4, 6                # [m², Pa, m⁴, m]
julia> Q, c, t_w = 323473.474 * (1/1e3)^3, 176.5e-3, 6.48e-3                # [m³, m, m]
julia> W(x) = -10e3(x) # [N/m]
julia> F_applied = ([(W, 1, 2, 3, 4)])
julia> x = 0:L:18
julia> elements = Dict{String, Tuple{Int, Int}}()
julia> nodeboundaryconditions = []
julia> for node ∈ 1:1:numberofelements+1
julia> if node == 1
julia> push!(nodeboundaryconditions, node)
julia> else
julia> elements["$(node-1)"] = (node-1, node)
julia> if x[node]%L == 0
julia> push!(nodeboundaryconditions, (node, (Inf, 0)))
julia> end
julia> end
julia> end
julia> preparedelements = prepareelements_beam(elements, E, I, L)
julia> sol = solve(preparedelements, nodeboundaryconditions, F_applied)     # [m]
julia> V = [sol.F[2i - 1] for i ∈ 1:1:length(sol.F)÷2]                      # [N]
julia> M = [sol.F[2i] for i ∈ 1:1:length(sol.F)÷2]                          # [N-m]
julia> maximum(shearstresses(elements, V, Q, t_w))                          # [Pa]
626638.3758201257

FiniteElementAnalysis.solve — Method

solve(elements::Dict{String, LineElements}, nodeboundaryconditions, nodeforces::Vector{<:Tuple{Integer, Vararg{<:Real}}}, dims::Integer=1) solve(elements::Dict{String, BeamElements}, nodeboundaryconditions, nodeforces::Vector{<:Tuple{Integer, Vararg{<:Real}}}, dims::Integer=2)

Perform Finite Element Analysis (FEA) with entries of elements and applied boundary conditions by the Direct Stiffness Method according to Hooke's Law of linear-elastic deformation.

If line elements, stiffness coefficient defined by: $k ≡ AE/L$ (typical of trusses with two-force members). If beam elements, stiffness coefficient defined by: $k ≡ EI/L³$.

See also: prepareelements_line, prepareelements_beam, globalstiffnessmatrix, and reducedglobalstiffnessmatrix.

Arguments

elements: the prepared dictionary according to element type.

nodeboundaryconditions: the tuple of nodes that either experience zero displacement (tuple item of integer for node) or prescribed displacement constrained by the stiffness of a certain element (tuple item of node integer and tuple of real displacement and integer element).

nodeforces::Vector{<:Tuple{Integer, Vararg{<:Real}}}: the vector of tuple pairs for nodes experiencing some applied force.

dims::Integer=1: the number of dimensions being analyzed.

alg=nothing: algorithm of choice to solve linear system of equations according to LinearSolve.jl.

Examples

1D

nodeboundaryconditions:= tuple of nodes with zero displacement.

julia> A, E, L = 4., 30.e6, 30. # [in², psi, in]
julia> elements = Dict("1"=>(1, 2), "2"=>(2, 3), "3"=>(3, 4))
julia> preparedelements = prepareelements_line(elements, A, E, L)
julia> nodeboundaryconditions = (1, 4)
julia> F_applied = [(2, 5.e3), (3, -10.e3)] # (node, force [lb])
julia> solve(preparedelements, nodeboundaryconditions, F_applied).U
4-element Vector{Float64}:
  0.0
  0.0
 -0.00125
  0.0

nodeboundaryconditions:= tuple of nodes with zero and prescribed displacements.

The tuple items are identified in the following way:

Node (1) with zero displacement.

Node (3) has a prescribed displacement of 25.e-3 constrained by the deformation of element 2.

julia> A, E, L = 4e-4, 210e9, 2 # [m², Pa, m]
julia> elements = Dict("1"=>(1, 2), "2"=>(2, 3))
julia> preparedelements = prepareelements_line(elements, A, E, L)
julia> nodeboundaryconditions = [1, (3, 25.e-3, 2)] # (node, displacement [m], element)
julia> F_applied = [(2, -5.e3)] # (node, force [N])
julia> solve(preparedelements, nodeboundaryconditions, F_applied).U
3-element Vector{Float64}:
 0.0
 0.012440476190476191
 0.025

2D

julia> A, E, L = 1, 10e6, 100 # [in², psi, in]
julia> angles = [120, 0, 210] # [degrees]
julia> L = L ./ abs.(cosd.(angles)) # [in]
julia> elements = Dict("1"=>(1, 2), "2"=>(1, 3), "3"=>(1, 4))
julia> preparedelements = prepareelements_line(elements, A, E, L, angles, dims=2)
julia> nodeboundaryconditions = [2, 3, 4]
julia> F_applied = [(1, (1e3, 1e3))] # (node, (force_x [lb], force_y [lb]))
julia> solve(preparedelements, nodeboundaryconditions, F_applied, dims=2).U
8-element Vector{Float64}:
 0.004226497308103743
 0.01577350269189626
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0